+++ android/src/main/java/jp/shiguredo/react/webrtckit/WebRTCRepository.java
@@ -4,6 +4,7 @@ import androidx.annotation.Nullable;
 import androidx.annotation.NonNull;
 import android.util.Pair;
 
+import org.webrtc.DataChannel;
 import org.webrtc.MediaStream;
 import org.webrtc.MediaStreamTrack;
 import org.webrtc.PeerConnection;
@@ -100,6 +101,7 @@ final class WebRTCRepository {
     //region RTP Sender
 
     final DualKeyMap<RtpSender> senders = new DualKeyMap<>();
+
     /**
      * Key is id, Value is associated stream ids.
      */
@@ -204,6 +206,31 @@ final class WebRTCRepository {
     //endregion
 
 
+    //region Data Channel
+
+    private final Map<String, DataChannel> dataChannelMap = new HashMap<>();
+
+    void addDataChannel(@NonNull final Pair<String, DataChannel> dataChannelPair) {
+        dataChannelMap.put(dataChannelPair.first, dataChannelPair.second);
+    }
+
+    void removeDataChannelByValueTag(@Nullable final String valueTag) {
+        if (valueTag == null) {
+            return;
+        }
+        dataChannelMap.remove(valueTag);
+    }
+
+    @Nullable
+    DataChannel getDataChannelByValueTag(@Nullable final String valueTag) {
+        if (valueTag == null) {
+            return null;
+        }
+        return dataChannelMap.get(valueTag);
+    }
+
+    //endregion
+
     /**
      * このリポジトリの中身を完全に空にします。
      * その際、格納されていたWebRTC関連のオブジェクトは、現在のところ、明示的に初期化されません。
@@ -223,6 +250,8 @@ final class WebRTCRepository {
         receiverStreamIdsMap.clear();
 
         transceivers.clear();
+
+        dataChannelMap.clear();
     }
 
     static final class DualKeyMap<V> {